Add files via upload
Problem Statement:
Write a program to test generalized hailstone numbers where xn = axn−1 +b if xn−1 is odd and xn = xn−1/2
if xn−1 is even where a and b are positive integers. For each of the a, b pairs less than or equal to 10, find
whether or not the sequence seems to converge. Once you can do that, how many different holding
patterns the sequence converges to. For instance, when a = 3 and b = 5 then 1 ! 8 ! 4 ! 2 ! 1 this
is holding pattern. However, if we start with a different number, say 5 we get a different holding pattern
5 ! 20 ! 10 ! 5.